requirements.txt
discord.py
yt-dlp
spotipy
ffmpeg-python
